Squashed commit of the following:

commit 1f5c49061072981d8d57aa3598a8e91e197de81d
Author: Alexander Ovechkin <ovov@yandex-team.ru>
Date:   Tue Sep 15 15:47:13 2020 +0300

    CLOUD-43368: added checkWithState

commit 69fa8890262c0f82be769a0c8205e56394006089
Author: Alexander Ovechkin <ovov@yandex-team.ru>
Date:   Mon Sep 14 03:04:26 2020 +0300

    CLOUD-43368: test fixes

commit 4c7701e63a8f2304197d4f9ba677218e8811c188
Author: Alexander Ovechkin <ovov@yandex-team.ru>
Date:   Mon Sep 14 02:08:11 2020 +0300

    CLOUD-43368: removed tmp debug logging

commit 3ddcfa26d9acb6fbe8634ab761fec9cc1e0092f6
Author: Alexander Ovechkin <ovov@yandex-team.ru>
Date:   Fri Sep 11 17:13:38 2020 +0300

    CLOUD-43368: using set for nlb listener representation in lb datasource

commit db181da50604ab1993deca76b33968434fd6a702
Author: Alexander Ovechkin <ovov@yandex-team.ru>
Date:   Fri Sep 11 02:02:29 2020 +0300

    CLOUD-43368:
    * fixed listener hash function
    * representing external/internal addresses as set instead of list

+++ yandex/base_lb_test.go
@@ -17,10 +17,10 @@ import (
 
 const lbDefaultNLBDescription = "nlb-descriprion"
 const lbDefaultListenerName = "nlb-listener"
-const lbDefaultListenerPort = 8080
+const lbDefaultListenerPort = int64(8080)
 const lbDefaultListenerProtocol = "tcp"
 const lbDefaultListenerIPVersion = "ipv4"
-const lbDefaultListenerTargetPort = 8080
+const lbDefaultListenerTargetPort = int64(8080)
 const lbDefaultHCHTTPName = "http"
 const lbDefaultHCHTTPInterval = 2
 const lbDefaultHCHTTPTimeout = 1
